pyproject.toml
[project]
name = "sql-assistant-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi==0.110.2",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "pandas==2.2.2",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "redis==5.0.3",
    "requests==2.31.0",
    "slowapi==0.1.8",
    "sql-metadata==2.13.0",
    "sqlalchemy==2.0.29",
    "sqlparse==0.5.0",
    "streamlit==1.36.0",
    "uvicorn==0.29.0",
]
